pass the path to work with
Now the program takes the 1st argument passed to it as the base path
from where the contents should be displayed.

The same is inturn passed to the Main window class and inturn used
by its update_lb function.

One can either pass a path to update_lb explicitly or else it will
automatically work with the curPath of the MainWindow instance.