Fix height of listbox, Try scroll to selected
Set the min and max height of scrolled window so that listbox
is inturn fixed to that height. So irrespective of if it has
less or more items, it will have this specified height.

Am trying to scroll to the selected row in listbox, but set_focus
doesnt seem to do the job. Have to think more.