btn press msgs dprints, Instead print related Content
Added a dprint function, One will have to set a debug level for the
message to be less than the global debug level, for it to print the
message.

Button press messages have been made into dprint, so they dont print
by default. Instead now they print the content that is selected /
operated on.