Rename to lastFile in config and handle appropriately
have renamed the config entry to lastFile from lastPath, to reflect
its actual content.

Also this requires one to extract the dirname from lastFile with path
in it and then suffix it to the program's base path.

Also take care of the case, where there may be no config file in the
base path passed to the program. (Well because its the first time
it is being run with that base path or else someone removed it).