Play on Double click
Row-Activated is called for <Enter> key press, as well as single
click on mouse button on a row in the list box. But I want to
play only on double click (and maybe enter key press). So had to
add button-release-event handler.

Also had to return False in the handler to get the 2nd mouse click
of a double click. >> Have to cross check this aspect once again
later <<

TODO: Currently i dont check if the double click is on a valid
row with data in the list box or in a empty region. Have to check
this later.

NOTE: First the 1st button release event occurs, then row-activated
occurs, then the 2nd button release event occurs during a double
click.