Stop wv media playback, new ev DocModel each time
stop_loading didnt help stop a already playing media in the webview,
so now I explicitly load a dummy internal page to stop the media.

As evince seems to need a new document model to be created for each
pdf loaded, so moved the creation of the same into play logic.