Take care of UI wrt Hide/UnHide of ListBox for now
As Up, Base and Last manipulate/change the directory / path location
of the program, which is inturn shown in the listbox to the left,
so if these buttons are pressed when the listbox is hidden, then
the program takes care of unhiding the listbox and then doing the
command.

NOTE: Without this handling resize runaway was also required, for
some reason, which I have think through bit more, while this avoids
this as well as keeps thinkgs logically in sync.