ListBox can scroll horizontally
Previously I was forcing the ListBox content to not be scrollable,
and thus inturn ListBox to expand along width to show its contents
fully.

This inturn was leading to the App having to increase its width.
However for directories with really long file names, this can lead
to the app window going beyond the screen width.

I trying to request the ListBox width to be with in reasonable limit
, if it tries to become too wide, doesnt seem to help.

So have now enabled horizontal scrolling, with this, the App size
remains fixed by default, unless user resizes it. Inturn the listbox
takes a fixed ~ 30% of total app width.