Put ListBox and a new Label inside a Grid
This is to allow a label below or above the listbox, which will
show has to which is the current directory whose contents it is
showing.

As this can be long name, so it is also embedded within the listbox
related scrolling window.

The actual full code for layout will be in a future commit. This
shows a partial examples of different ways of doing it.

The children of grid dont expand by default.