Move back to Grid and use it relatively better.
Moved listbox related label to top of listbox related left grid,
this ensures that even if we have a very huge list of contents,
the label will still be visible and directly at the top.

Use h and v expand as required hopefully.

NOTE: This 3 step commit and program in general, was and is to Keep
relearning things newly as and when required partially, when coming
back to things after long time gaps. And to scratch some immidiate
itch, So that some basic minimal reqd thing can be accomplished,
at some levels.