Evince Handled better
The previous flow where the EvinceView was reused, was crashing the
program in some cases. Maybe there is some cleanup I require to do
before reusing it.

Eitherway for now, I have taken the easy path and recreated the
EvinceView, when ever loading pdf file, and that fixes the issue
for now.